Levels now restart on a timer.
Managed to tweak level re-loading so that there is a brief pause before the actual restart that allows players to more easily process having just died. Also managed to make it so that the player's blaster has physics enabled when the player dies so that it properly falls out of their hand.

~George H