More or less finished the second level.
Not entirely sure about the fairness of the actual design of the second level, but I managed to get moving platforms working (and be able to assign multiple moving platforms to the same path), and get the level to a more refined state.

Also tweaked level loading delays so that the player does not advance to the next level if they die and kill the last enemy at around the same time (resulting in a double-KO).

~George H